extracted a custom hook to GET recipe data
