+++ src/meal-ui/src/dataHooks/defaultRequestOptions.js
@@ -1,0 +1,12 @@
+const defaultRequestOptions = Object.freeze({
+	mode: 'cors',
+	cache: 'no-cache',
+	credentials: 'same-origin',
+	headers: {
+		'Content-Type': 'application/json'
+	},
+	redirect: 'follow',
+	referrerPolicy: 'no-referrer'
+});
+
+export default defaultRequestOptions;
